genial-ai/requirements.txt
torch>=1.13.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
fastapi
uvicorn
python-multipart
